2.3.3 commented
In this part of the excercise the OpenValue should point to the column 3 (2018/"Open"). 
The excercise guide us to point twice to the column 6. Per startingPrice to endingPrice

So:
'        Find the starting price for the current ticker.
'Should be (j, 3) instead of (j, 6).  3 as this is the column for Open value and 6 is the column for Close value in 2018 worksheet.
    If Cells(j - 1, 1).Value <> ticker And Cells(j, 1).Value = ticker Then
        startingPrice = Cells(j, 3).Value
    End If